requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
matplotlib
torchcam